backend/requirements.txt
fastapi
uvicorn
pymongo
discord-webhook
pyyaml